SSCS-10865 / SSCS-11723 - Move code to correct handler, should be on adjourn case not issue.
